+++ lib/connection/connection.ex
@@ -38,8 +38,8 @@ defmodule RabbitMQStream.Connection do
   * `lazy` - If `true`, the connection won't starting until explicitly calling `connect/1`. Defaults to `false`.
 
 
-  # Subscribing to messages
-  You can subscribe to messages by calling `subscribe/5`:
+  # Consuming messages
+  You can consume messages by calling `subscribe/5`:
 
       {:ok, _subscription_id} = MyApp.MyConnection.subscribe("stream-01", self(), :next, 999)