+++ lib/consumer/consumer.ex
@@ -1,18 +1,18 @@
-defmodule RabbitMQStream.Subscriber do
+defmodule RabbitMQStream.Consumer do
   @moduledoc """
-  Used to declare a Persistent Subscriber module. It is able to process
+  Used to declare a Persistent Consumer module. It is able to process
   chunks by implementing the `handle_chunk/1` or `handle_chunk/2` callbacks.
 
   # Usage
 
-      defmodule MyApp.MySubscriber do
-        use RabbitMQStream.Subscriber,
+      defmodule MyApp.MyConsumer do
+        use RabbitMQStream.Consumer,
           connection: MyApp.MyConnection,
           stream_name: "my_stream",
           initial_offset: :first
 
         @impl true
-        def handle_chunk(%RabbitMQStream.OsirisChunk{} = _chunk, _subscriber) do
+        def handle_chunk(%RabbitMQStream.OsirisChunk{} = _chunk, _consumer) do
           :ok
         end
       end
@@ -21,14 +21,14 @@ defmodule RabbitMQStream.Subscriber do
   # Parameters
 
   * `:connection` - The connection module to use. This is required.
-  * `:stream_name` - The name of the stream to subscribe to. This is required.
-  * `:initial_offset` - The initial offset to subscribe from. This is required.
+  * `:stream_name` - The name of the stream to consume. This is required.
+  * `:initial_offset` - The initial offset. This is required.
   * `:initial_credit` - The initial credit to request from the server. Defaults to `50_000`.
   * `:offset_tracking` - Offset tracking strategies to use. Defaults to `[count: [store_after: 50]]`.
   * `:flow_control` - Flow control strategy to use. Defaults to `[count: [credit_after: {:count, 1}]]`.
   * `:private` - Private data that can hold any value, and is passed to the `handle_chunk/2` callback.
   * `:serializer` - The module to use to decode the message. Defaults to `__MODULE__`,
-    which means that the subscriber will use the `decode!/1` callback to decode the message, which is implemented by default to return the message as is.
+    which means that the consumer will use the `decode!/1` callback to decode the message, which is implemented by default to return the message as is.
 
   * `:properties` - Define the properties of the subscription. Can only have one option at a time.
     * `:single_active_consumer`: set to `true` to enable [single active consumer](https://blog.rabbitmq.com/posts/2022/07/rabbitmq-3-11-feature-preview-single-active-consumer-for-streams/) for this subscription.
@@ -39,47 +39,47 @@ defmodule RabbitMQStream.Subscriber do
 
   # Offset Tracking
 
-  The subscriber is able to track its progress in the stream by storing its
+  The consumer is able to track its progress in the stream by storing its
   latests offset in the stream. Check [Offset Tracking with RabbitMQ Streams(https://blog.rabbitmq.com/posts/2021/09/rabbitmq-streams-offset-tracking/) for more information on
   how offset tracking works.
 
-  The subscriber can be configured to use different offset tracking strategies,
+  The consumer can be configured to use different offset tracking strategies,
   which decide when to store the offset in the stream. You can implement your
-  own strategy by implementing the `RabbitMQStream.Subscriber.OffsetTracking.Strategy`
+  own strategy by implementing the `RabbitMQStream.Consumer.OffsetTracking.Strategy`
   behaviour, and passing it to the `:offset_tracking` option. It defaults to
-  `RabbitMQStream.Subscriber.OffsetTracking.CountStrategy`, which stores the
+  `RabbitMQStream.Consumer.OffsetTracking.CountStrategy`, which stores the
   offset after, by default, every 50_000 messages.
 
   # Flow Control
 
-  The RabbitMQ Streams server requires that the subscriber declares how many
+  The RabbitMQ Streams server requires that the consumer declares how many
   messages it is able to process at a time. This is done by informing an amount
   of 'credits' to the server. After every chunk is sent, one credit is consumed,
   and the server will send messages only if there are credits available.
 
-  We can configure the subscriber to automatically request more credits based on
-  a strategy. By default it uses the `RabbitMQStream.Subscriber.FlowControl.MessageCount`,
+  We can configure the consumer to automatically request more credits based on
+  a strategy. By default it uses the `RabbitMQStream.Consumer.FlowControl.MessageCount`,
   which requests 1 additional credit for every 1 processed chunk. Please check
-  the RabbitMQStream.Subscriber.FlowControl.MessageCount module for more information.
+  the RabbitMQStream.Consumer.FlowControl.MessageCount module for more information.
 
-  You can also call `RabbitMQStream.Subscriber.credit/2` to manually add more
+  You can also call `RabbitMQStream.Consumer.credit/2` to manually add more
   credits to the subscription, or implement your own strategy by implementing
-  the `RabbitMQStream.Subscriber.FlowControl.Strategy` behaviour, and passing
+  the `RabbitMQStream.Consumer.FlowControl.Strategy` behaviour, and passing
   it to the `:flow_control` option.
 
   You can find more information on the [RabbitMQ Streams documentation](https://www.rabbitmq.com/stream.html#flow-control).
 
   If you want an external process to be fully in control of the flow control
-  of a subscriber, you can set the `:flow_control` option to `false`. Then
-  you can call `RabbitMQStream.Subscriber.credit/2` to manually add more
+  of a consumer, you can set the `:flow_control` option to `false`. Then
+  you can call `RabbitMQStream.Consumer.credit/2` to manually add more
   credits to the subscription.
 
 
   # Configuration
 
-  You can configure each subscriber with:
+  You can configure each consumer with:
 
-      config :rabbitmq_stream, MyApp.MySubscriber,
+      config :rabbitmq_stream, MyApp.MyConsumer,
         connection: MyApp.MyConnection,
         stream_name: "my_stream",
         initial_offset: :first,
@@ -91,16 +91,16 @@ defmodule RabbitMQStream.Subscriber do
   These options are overriden by the options passed to the `use` macro, which
   are overriden by the options passed to `start_link/1`.
 
-  And also you can override the defaults of all subscribers with:
+  And also you can override the defaults of all consumers with:
 
         config :rabbitmq_stream, :defaults,
-          subscribers: [
+          consumers: [
             connection: MyApp.MyConnection,
             initial_credit: 50_000,
             # ...
           ],
 
-  Globally configuring all subscribers ignores the following options:
+  Globally configuring all consumers ignores the following options:
 
   * `:stream_name`
   * `:offset_reference`
@@ -110,8 +110,8 @@ defmodule RabbitMQStream.Subscriber do
   defmacro __using__(opts) do
     quote bind_quoted: [opts: opts], location: :keep do
       use GenServer
-      @behaviour RabbitMQStream.Subscriber
-      alias RabbitMQStream.Subscriber.{FlowControl, OffsetTracking}
+      @behaviour RabbitMQStream.Consumer
+      alias RabbitMQStream.Consumer.{FlowControl, OffsetTracking}
       alias RabbitMQStream.Message.Request
 
       @opts opts
@@ -127,7 +127,7 @@ defmodule RabbitMQStream.Subscriber do
       def start_link(opts \\ []) do
         opts =
           Application.get_env(:rabbitmq_stream, :defaults, [])
-          |> Keyword.get(:subscribers, [])
+          |> Keyword.get(:consumers, [])
           |> Keyword.drop([:stream_name, :offset_reference, :private])
           |> Keyword.merge(Application.get_env(:rabbitmq_stream, :defaults, []) |> Keyword.take([:serializer]))
           |> Keyword.merge(Application.get_env(:rabbitmq_stream, __MODULE__, []))
@@ -160,7 +160,7 @@ defmodule RabbitMQStream.Subscriber do
           |> Keyword.put(:offset_tracking, OffsetTracking.Strategy.init(opts[:offset_tracking], opts))
           |> Keyword.put(:flow_control, FlowControl.Strategy.init(opts[:flow_control], opts))
 
-        state = struct(RabbitMQStream.Subscriber, opts)
+        state = struct(RabbitMQStream.Consumer, opts)
 
         {:ok, state, {:continue, {:init, opts}}}
       end
@@ -251,7 +251,7 @@ defmodule RabbitMQStream.Subscriber do
         if function_exported?(__MODULE__, :handle_update, 2) do
           case apply(__MODULE__, :handle_update, [state, request.data.active]) do
             {:ok, offset} ->
-              Logger.debug("Subscriber upgraded to active consumer")
+              Logger.debug("Consumer upgraded to active consumer")
               state.connection.respond(request, offset: offset, code: :ok)
               {:noreply, state}
 
@@ -281,7 +281,7 @@ defmodule RabbitMQStream.Subscriber do
 
       def decode!(message), do: message
 
-      defoverridable RabbitMQStream.Subscriber
+      defoverridable RabbitMQStream.Consumer
     end
   end
 
@@ -295,7 +295,7 @@ defmodule RabbitMQStream.Subscriber do
     chunk received.
 
     Optionally if you implement `handle_chunk/2`, it also passes the current
-    state of the subscriber. It can be used to access the `private` field
+    state of the consumer. It can be used to access the `private` field
     passed to `start_link/1`, and other fields.
 
     The return value is ignored.
@@ -303,7 +303,7 @@ defmodule RabbitMQStream.Subscriber do
   @callback handle_chunk(chunk :: RabbitMQStream.OsirisChunk.t()) :: term()
   @callback handle_chunk(chunk :: RabbitMQStream.OsirisChunk.t(), state :: t()) :: term()
 
-  @callback handle_update(subscriber :: t(), flag :: boolean()) ::
+  @callback handle_update(consumer :: t(), flag :: boolean()) ::
               {:ok, RabbitMQStream.Connection.offset()} | {:error, any()}
 
   @callback decode!(message :: String.t()) :: term()
@@ -332,27 +332,27 @@ defmodule RabbitMQStream.Subscriber do
           connection: RabbitMQStream.Connection.t(),
           stream_name: String.t(),
           id: non_neg_integer() | nil,
-          offset_tracking: [{RabbitMQStream.Subscriber.OffsetTracking.Strategy.t(), term()}],
-          flow_control: {RabbitMQStream.Subscriber.FlowControl.Strategy.t(), term()},
+          offset_tracking: [{RabbitMQStream.Consumer.OffsetTracking.Strategy.t(), term()}],
+          flow_control: {RabbitMQStream.Consumer.FlowControl.Strategy.t(), term()},
           last_offset: non_neg_integer() | nil,
           private: any(),
           credits: non_neg_integer(),
           initial_credit: non_neg_integer(),
           serializer: {module(), atom()} | (String.t() -> term()) | nil,
-          properties: [RabbitMQStream.Message.Types.SubscribeRequestData.property()]
+          properties: [RabbitMQStream.Message.Types.ConsumerequestData.property()]
         }
 
-  @type subscriber_option ::
+  @type consumer_option ::
           {:offset_reference, String.t()}
           | {:connection, RabbitMQStream.Connection.t()}
           | {:stream_name, String.t()}
           | {:initial_offset, RabbitMQStream.Connection.offset()}
           | {:initial_credit, non_neg_integer()}
-          | {:offset_tracking, [{RabbitMQStream.Subscriber.OffsetTracking.Strategy.t(), term()}]}
-          | {:flow_control, {RabbitMQStream.Subscriber.FlowControl.Strategy.t(), term()}}
+          | {:offset_tracking, [{RabbitMQStream.Consumer.OffsetTracking.Strategy.t(), term()}]}
+          | {:flow_control, {RabbitMQStream.Consumer.FlowControl.Strategy.t(), term()}}
           | {:private, any()}
           | {:serializer, {module(), atom()} | (String.t() -> term())}
-          | {:properties, [RabbitMQStream.Message.Types.SubscribeRequestData.property()]}
+          | {:properties, [RabbitMQStream.Message.Types.ConsumerequestData.property()]}
 
-  @type opts :: [subscriber_option()]
+  @type opts :: [consumer_option()]
 end